Refactors Sutra UI elements to separate files Refactors Sutra property type input to Components selector
